+++ finance/src/test/kotlin/net/corda/contracts/asset/ObligationTests.kt
@@ -13,6 +13,7 @@ import net.corda.core.utilities.NonEmptySet
 import net.corda.core.utilities.OpaqueBytes
 import net.corda.core.utilities.days
 import net.corda.core.utilities.hours
+import net.corda.finance.*
 import net.corda.testing.*
 import net.corda.testing.contracts.DummyState
 import net.corda.testing.node.MockServices
@@ -44,7 +45,7 @@ class ObligationTests {
             quantity = 1000.DOLLARS.quantity,
             beneficiary = CHARLIE
     )
-    val outState = inState.copy(beneficiary = AnonymousParty(DUMMY_PUBKEY_2))
+    val outState = inState.copy(beneficiary = AnonymousParty(BOB_PUBKEY))
     val miniCorpServices = MockServices(MINI_CORP_KEY)
     val notaryServices = MockServices(DUMMY_NOTARY_KEY)
 
@@ -68,20 +69,21 @@ class ObligationTests {
     fun trivial() {
         transaction {
             input { inState }
-            this `fails with` "the amounts balance"
 
             tweak {
                 output { outState.copy(quantity = 2000.DOLLARS.quantity) }
+                command(CHARLIE.owningKey) { Obligation.Commands.Move() }
                 this `fails with` "the amounts balance"
             }
             tweak {
                 output { outState }
-                // No command arguments
+                command(CHARLIE.owningKey) { DummyCommandData }
+                // Invalid command
                 this `fails with` "required net.corda.contracts.asset.Obligation.Commands.Move command"
             }
             tweak {
                 output { outState }
-                command(DUMMY_PUBKEY_2) { Obligation.Commands.Move() }
+                command(BOB_PUBKEY) { Obligation.Commands.Move() }
                 this `fails with` "the owning keys are a subset of the signing keys"
             }
             tweak {
@@ -224,8 +226,8 @@ class ObligationTests {
     @Test
     fun `generate close-out net transaction`() {
         initialiseTestSerialization()
-        val obligationAliceToBob = oneMillionDollars.OBLIGATION between Pair(ALICE, BOB)
-        val obligationBobToAlice = oneMillionDollars.OBLIGATION between Pair(BOB, ALICE)
+        val obligationAliceToBob = getStateAndRef(oneMillionDollars.OBLIGATION between Pair(ALICE, BOB))
+        val obligationBobToAlice = getStateAndRef(oneMillionDollars.OBLIGATION between Pair(BOB, ALICE))
         val tx = TransactionBuilder(DUMMY_NOTARY).apply {
             Obligation<Currency>().generateCloseOutNetting(this, ALICE, obligationAliceToBob, obligationBobToAlice)
         }.toWireTransaction()
@@ -236,8 +238,8 @@ class ObligationTests {
     @Test
     fun `generate close-out net transaction with remainder`() {
         initialiseTestSerialization()
-        val obligationAliceToBob = (2000000.DOLLARS `issued by` defaultIssuer).OBLIGATION between Pair(ALICE, BOB)
-        val obligationBobToAlice = oneMillionDollars.OBLIGATION between Pair(BOB, ALICE)
+        val obligationAliceToBob = getStateAndRef((2000000.DOLLARS `issued by` defaultIssuer).OBLIGATION between Pair(ALICE, BOB))
+        val obligationBobToAlice = getStateAndRef(oneMillionDollars.OBLIGATION between Pair(BOB, ALICE))
         val tx = TransactionBuilder(DUMMY_NOTARY).apply {
             Obligation<Currency>().generateCloseOutNetting(this, ALICE, obligationAliceToBob, obligationBobToAlice)
         }.toWireTransaction()
@@ -251,10 +253,10 @@ class ObligationTests {
     @Test
     fun `generate payment net transaction`() {
         initialiseTestSerialization()
-        val obligationAliceToBob = oneMillionDollars.OBLIGATION between Pair(ALICE, BOB)
-        val obligationBobToAlice = oneMillionDollars.OBLIGATION between Pair(BOB, ALICE)
+        val obligationAliceToBob = getStateAndRef(oneMillionDollars.OBLIGATION between Pair(ALICE, BOB))
+        val obligationBobToAlice = getStateAndRef(oneMillionDollars.OBLIGATION between Pair(BOB, ALICE))
         val tx = TransactionBuilder(DUMMY_NOTARY).apply {
-            Obligation<Currency>().generatePaymentNetting(this, obligationAliceToBob.amount.token, DUMMY_NOTARY, obligationAliceToBob, obligationBobToAlice)
+            Obligation<Currency>().generatePaymentNetting(this, obligationAliceToBob.state.data.amount.token, DUMMY_NOTARY, obligationAliceToBob, obligationBobToAlice)
         }.toWireTransaction()
         assertEquals(0, tx.outputs.size)
     }
@@ -263,15 +265,23 @@ class ObligationTests {
     @Test
     fun `generate payment net transaction with remainder`() {
         initialiseTestSerialization()
-        val obligationAliceToBob = oneMillionDollars.OBLIGATION between Pair(ALICE, BOB)
-        val obligationBobToAlice = (2000000.DOLLARS `issued by` defaultIssuer).OBLIGATION between Pair(BOB, ALICE)
-        val tx = TransactionBuilder(null).apply {
-            Obligation<Currency>().generatePaymentNetting(this, obligationAliceToBob.amount.token, DUMMY_NOTARY, obligationAliceToBob, obligationBobToAlice)
+        val obligationAliceToBob = getStateAndRef(oneMillionDollars.OBLIGATION between Pair(ALICE, BOB))
+        val obligationAliceToBobState = obligationAliceToBob.state.data
+        val obligationBobToAlice = getStateAndRef((2000000.DOLLARS `issued by` defaultIssuer).OBLIGATION between Pair(BOB, ALICE))
+        val obligationBobToAliceState = obligationBobToAlice.state.data
+        val tx = TransactionBuilder(DUMMY_NOTARY).apply {
+            Obligation<Currency>().generatePaymentNetting(this, obligationAliceToBobState.amount.token, DUMMY_NOTARY, obligationAliceToBob, obligationBobToAlice)
         }.toWireTransaction()
         assertEquals(1, tx.outputs.size)
-        val expected = obligationBobToAlice.copy(quantity = obligationBobToAlice.quantity - obligationAliceToBob.quantity)
+        val expected = obligationBobToAliceState.copy(quantity = obligationBobToAliceState.quantity - obligationAliceToBobState.quantity)
         val actual = tx.getOutput(0)
         assertEquals(expected, actual)
+    }
+
+    private inline fun <reified T: ContractState> getStateAndRef(state: T): StateAndRef<T> {
+        val txState = TransactionState(state, DUMMY_NOTARY)
+        return StateAndRef(txState, StateRef(SecureHash.randomSHA256(), 0))
+
     }
 
     /** Test generating a transaction to mark outputs as having defaulted. */
@@ -596,15 +606,20 @@ class ObligationTests {
     @Test
     fun zeroSizedValues() {
         transaction {
-            input { inState }
-            input { inState.copy(quantity = 0L) }
-            this `fails with` "zero sized inputs"
-        }
-        transaction {
-            input { inState }
-            output { inState }
-            output { inState.copy(quantity = 0L) }
-            this `fails with` "zero sized outputs"
+            command(CHARLIE.owningKey) { Obligation.Commands.Move() }
+            tweak {
+                input { inState }
+                input { inState.copy(quantity = 0L) }
+
+                this `fails with` "zero sized inputs"
+            }
+            tweak {
+                input { inState }
+                output { inState }
+                output { inState.copy(quantity = 0L) }
+
+                this `fails with` "zero sized outputs"
+            }
         }
     }
 
@@ -614,6 +629,7 @@ class ObligationTests {
         transaction {
             input { inState }
             output { outState `issued by` MINI_CORP }
+            command(MINI_CORP_PUBKEY) { Obligation.Commands.Move() }
             this `fails with` "the amounts balance"
         }
         // Can't mix currencies.
@@ -621,6 +637,7 @@ class ObligationTests {
             input { inState }
             output { outState.copy(quantity = 80000, template = megaCorpDollarSettlement) }
             output { outState.copy(quantity = 20000, template = megaCorpPoundSettlement) }
+            command(MINI_CORP_PUBKEY) { Obligation.Commands.Move() }
             this `fails with` "the amounts balance"
         }
         transaction {
@@ -629,10 +646,11 @@ class ObligationTests {
                 inState.copy(
                         quantity = 15000,
                         template = megaCorpPoundSettlement,
-                        beneficiary = AnonymousParty(DUMMY_PUBKEY_2)
+                        beneficiary = AnonymousParty(BOB_PUBKEY)
                 )
             }
             output { outState.copy(quantity = 115000) }
+            command(MINI_CORP_PUBKEY) { Obligation.Commands.Move() }
             this `fails with` "the amounts balance"
         }
         // Can't have superfluous input states from different issuers.
@@ -702,19 +720,21 @@ class ObligationTests {
 
             // Can't merge them together.
             tweak {
-                output { inState.copy(beneficiary = AnonymousParty(DUMMY_PUBKEY_2), quantity = 200000L) }
+                output { inState.copy(beneficiary = AnonymousParty(BOB_PUBKEY), quantity = 200000L) }
+                command(CHARLIE.owningKey) { Obligation.Commands.Move() }
                 this `fails with` "the amounts balance"
             }
             // Missing MiniCorp deposit
             tweak {
-                output { inState.copy(beneficiary = AnonymousParty(DUMMY_PUBKEY_2)) }
-                output { inState.copy(beneficiary = AnonymousParty(DUMMY_PUBKEY_2)) }
+                output { inState.copy(beneficiary = AnonymousParty(BOB_PUBKEY)) }
+                output { inState.copy(beneficiary = AnonymousParty(BOB_PUBKEY)) }
+                command(CHARLIE.owningKey) { Obligation.Commands.Move() }
                 this `fails with` "the amounts balance"
             }
 
             // This works.
-            output { inState.copy(beneficiary = AnonymousParty(DUMMY_PUBKEY_2)) }
-            output { inState.copy(beneficiary = AnonymousParty(DUMMY_PUBKEY_2)) `issued by` MINI_CORP }
+            output { inState.copy(beneficiary = AnonymousParty(BOB_PUBKEY)) }
+            output { inState.copy(beneficiary = AnonymousParty(BOB_PUBKEY)) `issued by` MINI_CORP }
             command(CHARLIE.owningKey) { Obligation.Commands.Move() }
             this.verifies()
         }
@@ -724,12 +744,12 @@ class ObligationTests {
     fun multiCurrency() {
         // Check we can do an atomic currency trade tx.
         transaction {
-            val pounds = Obligation.State(Lifecycle.NORMAL, MINI_CORP, megaCorpPoundSettlement, 658.POUNDS.quantity, AnonymousParty(DUMMY_PUBKEY_2))
+            val pounds = Obligation.State(Lifecycle.NORMAL, MINI_CORP, megaCorpPoundSettlement, 658.POUNDS.quantity, AnonymousParty(BOB_PUBKEY))
             input { inState `owned by` CHARLIE }
             input { pounds }
-            output { inState `owned by` AnonymousParty(DUMMY_PUBKEY_2) }
+            output { inState `owned by` AnonymousParty(BOB_PUBKEY) }
             output { pounds `owned by` CHARLIE }
-            command(CHARLIE.owningKey, DUMMY_PUBKEY_2) { Obligation.Commands.Move() }
+            command(CHARLIE.owningKey, BOB_PUBKEY) { Obligation.Commands.Move() }
 
             this.verifies()
         }