+++ node/src/test/kotlin/net/corda/node/InteractiveShellTest.kt
@@ -4,10 +4,7 @@ import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
 import net.corda.core.concurrent.CordaFuture
 import net.corda.core.contracts.Amount
 import net.corda.core.crypto.SecureHash
-import net.corda.core.flows.FlowContext
-import net.corda.core.flows.FlowInitiator
-import net.corda.core.flows.FlowLogic
-import net.corda.core.flows.StateMachineRunId
+import net.corda.core.flows.*
 import net.corda.core.identity.Party
 import net.corda.core.internal.FlowStateMachine
 import net.corda.core.node.ServiceHub
@@ -32,6 +29,7 @@ class InteractiveShellTest {
         constructor(amount: Amount<Currency>) : this(amount.toString())
         constructor(pair: Pair<Amount<Currency>, SecureHash.SHA256>) : this(pair.toString())
         constructor(party: Party) : this(party.name.toString())
+
         override fun call() = a
     }
 
@@ -91,13 +89,9 @@ class InteractiveShellTest {
         override val id: StateMachineRunId get() = throw UnsupportedOperationException()
         override val resultFuture: CordaFuture<Any?> get() = throw UnsupportedOperationException()
         override val flowInitiator: FlowInitiator get() = throw UnsupportedOperationException()
-
-        override fun checkFlowPermission(permissionName: String, extraAuditData: Map<String, String>) {
-            // Do nothing
-        }
-
-        override fun recordAuditEvent(eventType: String, comment: String, extraAuditData: Map<String, String>) {
-            // Do nothing
-        }
+        override fun checkFlowPermission(permissionName: String, extraAuditData: Map<String, String>) = Unit
+        override fun recordAuditEvent(eventType: String, comment: String, extraAuditData: Map<String, String>) = Unit
+        override fun flowStackSnapshot(flowClass: Class<out FlowLogic<*>>): FlowStackSnapshot? = null
+        override fun persistFlowStackSnapshot(flowClass: Class<out FlowLogic<*>>) = Unit
     }
 }