+++ docs/source/release-notes.rst
@@ -6,6 +6,10 @@ Here are release notes for each snapshot release from M9 onwards.
 Unreleased
 ----------
 
+* Merged handling of well known and confidential identities in the identity service.
+
+* Remove `IssuerFlow` as it allowed nodes to request arbitrary amounts of cash to be issued from any remote node.
+
 Milestone 14
 ------------